Add ability to load remote files as a release values
Feature was requested in issues: [one][1], [two][2]

As a base logic I took an example with env values loader. To achieve that I added remote loader to the state on a parse step, in that case we have all needed information to load remote files in the same way as we are doing with env values.

[1]: https://github.com/roboll/helmfile/issues/1838
[2]: https://github.com/roboll/helmfile/issues/1499

Signed-off-by: Vladimir Kuznichenkov <kuzaxak.tech@gmail.com>