Enhance `make fmt` to cover running gci for golangci-lint compat (#515)
* Enhance `make fmt` to cover running gci for golangci-lint compat

I have been using `make fmt` for formatting code before submitting a pull request. It turned out not to work in some cases where you added a new import because the default go-fmt obviously do not organize imports as the golangci-lint and gci expect.
This fixes that, by adding a gci command to the `make fmt` target.
Please note that this does not cover installing gci. If you need, please submit another pull request to add something like a `make install-gci` and make the `make fmt` depenedent on that target, so that one can get automatic gci installation when running `make fmt` for first time.

Signed-off-by: Yusuke Kuoka <ykuoka@gmail.com>

* We don't need running gci via make on CI as we already run it as part of golangci-lint run

Signed-off-by: Yusuke Kuoka <ykuoka@gmail.com>

Signed-off-by: Yusuke Kuoka <ykuoka@gmail.com>