feat: Add missingFileHandlerConfig.ignoreMissingGitBranch
This enables you to specify missingFileHandlerConfig along with the existing missingFileHandler per release and per environment to fine-tune how the missing file handler works.

The first and the only supported setting is `ignoreMissingGitBranch,` which makes the handler treat Git command errors like `pathspec 'develop' did not match any file(s) known to git` as non-errors, so that missing branches are handled same as missing files.

Resolves #392

Signed-off-by: Yusuke Kuoka <ykuoka@gmail.com>