Relax the forbid-env-with-releases policy
This relaxes the policy proposed in v1 proposal and implemented via #592 and #683 so that it does not unnecessary fail for v1 with the plain yaml helmfile.yaml file.

The policy is there to migrate away from the legacy double-rendering feature that is involved when the helmfile.yaml is a go template. helmfile.yaml and helmfile.yaml.gotmpl are both go templates in v0. However, helmfile.yaml is not a go template in v1. That is why we can justify skipping the policy checker for helmfile.yaml on v1.

Follow-up for #592 and #683

Signed-off-by: Yusuke Kuoka <ykuoka@gmail.com>