Still attempt upgrades on non-deployed releases
Currently, if _any_ helm release is extant in the cluster at all,
"helmfile apply" will attempt to run a "helm diff" between the
release in the cluster and the proposed release, and only run helm
upgrade if the diff shows a change.

While this is a sensible check for "deployed" releases, for failed
(and possibly any other) release states, this results in the worst
possible outcome: helmfile *appears* to finish successfully, while
completely ignoring the failed release.

This skips the diff for non-deployed releases, treating it in the
same fashion as non-installed releases, always running the upgrade
for these particular releases. This way, if re-running with the
same chart would succeed, we have the opportunity to fix things on
rerun; if it wouldn't, by rerunning the upgrade, we will again see
the error rather than it going under the radar.

Signed-off-by: Joaquin Lopez <mrgus@disco-zombie.net>