Speed up isReleaseInstalled
We are facing an issue where we have tens (up to 100) of releases, and around 20% of those are `installed: false`.

The `isReleaseInstalled` check makes applying chart upgrades extremely slow for us, because `helm list ...`  turns out to be the a really slow operation in Helm.

Given that `helm list ...` is invoked for each individual release, we can use a more performant endpoint like `helm status ...` here.

Signed-off-by: Nikita Nefedov <inefedor@gmail.com>