chore: Remove unnecessary field from StateCreator
Although it was a public field, it has been true by default and no code turns it off.

Let's remove the field entirely, so that it's theoretically a little bit easier to understand the code.

If any third-party code is using this field, it will break, but it's unlikely-
If you ever turned it off, you are likely to have a helmfile.yaml that can be parsed only with Strict=false,
which is impossible to be parsed with the standard Helmfile; a portability issue!

Signed-off-by: Yusuke Kuoka <ykuoka@gmail.com>