feat: Ensure repo update is only run once (#1310)
By default `helm dep build` will fetch repo updates before building the
chart. For helmfile this means for every chart with remote dependencies,
another repo update is performed. This isn't necessary as the cache only
needs to be refreshed once.

By performing a `helm repo update` before running any `helm dep`, it is
safe to pass `--skip-refresh` to all `helm dep build` commands as the
cache will be up-to-date.

A few notes to be aware of for this change:
* If there are **no remote dependencies** then this will lead to an
  extra refresh that could be unnecessary. However, a single repo update
  is usually quite fast and can still be skipped with a manual
  `--skip-refresh`.
* There are no tests for this new behaviour yet
* I think even this single update might be unnecessary as any initial
  `helm repo add --force-update` that is normally performed anyway and
  should already refresh all repos. However, I wasn't confident enough
  to simply force this flag at all times and considered this the safer
  option.

Resolves #1310

Signed-off-by: javex <code@inexplicity.de>